move BulkResponse to common
