+++ core/src/main/java/com/netflix/conductor/service/WorkflowBulkServiceImpl.java
@@ -15,7 +15,7 @@ package com.netflix.conductor.service;
 import com.netflix.conductor.annotations.Audit;
 import com.netflix.conductor.annotations.Trace;
 import com.netflix.conductor.core.execution.WorkflowExecutor;
-import com.netflix.conductor.core.utils.BulkResponse;
+import com.netflix.conductor.common.model.BulkResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
